:white_check_mark:test : 일치하는 로또 개수를 구한다
